+++ common/src/results.rs
@@ -2,12 +2,12 @@ use crate::question::Question;
 
 #[derive(Debug)]
 pub struct BestThresholdResult {
-    pub loss: f32,
-    pub threshold_value: f32,
+    pub loss: f64,
+    pub threshold_value: f64,
 }
 
 #[derive(Debug)]
 pub struct BestSplitResult {
-    pub gain: f32,
+    pub gain: f64,
     pub question: Question,
 }
